[COMMON] gpu: g2d: fix implicit type conversion
Prevent CID: 202021
Prevent CID: 202003
Prevent CID: 202000
Prevent CID: 201961
Prevent CID: 201909
Prevent CID: 201887
Prevent CID: 201224

(cherry picked from commit a6bc6c9c652eb2eefc3f10f2dc07bd4f427e291f)
Bug: 165562344

+++ drivers/gpu/exynos/g2d/g2d_drv.c
@@ -544,7 +544,7 @@ static long g2d_compat_ioctl(struct file *filp,
 
 	switch (cmd) {
 	case COMPAT_G2D_IOC_PROCESS:
-		cmd = G2D_IOC_PROCESS;
+		cmd = (unsigned int)G2D_IOC_PROCESS;
 		break;
 	case G2D_IOC_PRIORITY:
 	case G2D_IOC_PERFORMANCE:
